[VL-1] GenreDao and test are complete

+++ video-library-adapter/jdbc-adapter/src/main/java/it/proconsole/library/video/adapter/jdbc/repository/dao/FilmGenreDao.java
@@ -19,7 +19,7 @@ public class FilmGenreDao extends DatabaseDao<FilmGenreEntity> {
   RowMapper<FilmGenreEntity> rowMapper() {
     return (rs, rowNum) -> new FilmGenreEntity(
         rs.getLong("film_id"),
-        rs.getInt("genre_id")
+        rs.getLong("genre_id")
     );
   }
 }
