+++ video-library-adapter/jdbc-adapter/src/main/java/it/proconsole/library/video/adapter/jdbc/repository/entity/FilmGenreEntity.java
@@ -1,6 +1,6 @@
 package it.proconsole.library.video.adapter.jdbc.repository.entity;
 
-public record FilmGenreEntity(long filmId, int genreId) implements EntityWithId {
+public record FilmGenreEntity(Long filmId, Long genreId) implements EntityWithId {
   @Override
   public Long id() {
     return filmId;
